# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethree API",
    "version": "1.0.0",
    "description": "API for Ethree - Eat, Enjoy, Entertainment platform (Supabase PostgreSQL Version)"
  },
  "servers": [
    {
      "url": "https://e3-e4-backend.vercel.app",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/send-otp": {
      "post": {
        "summary": "Send OTP to mobile number",
        "description": "Sends a 6-digit OTP via SMS (AUTHKEY). In development mode (no AUTHKEY env var),\nthe OTP is returned in the response as `debugOtp`.\nThe `location` (E3 or E4) is saved alongside the OTP so that `/verify-otp`\ndoes not need to ask for it again.\n",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mobile"
                ],
                "properties": {
                  "mobile": {
                    "type": "string",
                    "example": "9876543210",
                    "description": "10-digit Indian mobile number"
                  },
                  "location": {
                    "type": "string",
                    "enum": [
                      "E3",
                      "E4"
                    ],
                    "default": "E3",
                    "description": "Park location — saved in the OTP record and auto-embedded in the JWT on verification"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP sent successfully"
                    },
                    "mobile": {
                      "type": "string",
                      "example": "9876543210"
                    },
                    "debugOtp": {
                      "type": "string",
                      "example": "123456",
                      "description": "Only returned in development mode (when AUTHKEY env var is not set)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing or invalid mobile)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error or SMS provider failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP and login / register",
        "description": "Verifies the submitted OTP against the stored value.\n- If the user **exists** → logs them in and returns tokens.\n- If the user **does not exist** → creates a new account with role `customer` and returns tokens.\nThe **location (E3/E4) is read automatically from the OTP record** — you do not need to send it again.\nIt is then embedded into the JWT as the `type` claim for all future requests.\nA long-lived `refreshToken` is also set as an **HTTP-Only cookie** (not visible in the response body).\n",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mobile",
                  "otp",
                  "location"
                ],
                "properties": {
                  "mobile": {
                    "type": "string",
                    "example": "9876543210"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456"
                  },
                  "location": {
                    "type": "string",
                    "example": "E3"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login / registration successful",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-Only refreshToken cookie (30 days)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Short-lived JWT access token (15 minutes) — contains id, role, type (location)"
                    },
                    "isNewUser": {
                      "type": "boolean",
                      "description": "true if the account was just created"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP, expired OTP, or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "summary": "Refresh Access Token",
        "description": "Uses the **`refreshToken` HTTP-Only cookie** (set automatically by `/verify-otp`) to issue a new short-lived access token.\nNo request body is needed — the cookie is sent automatically by the browser.\nIf the cookie is missing or invalid, returns 401/403.\n",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "New access token issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "New JWT access token (valid for 15 minutes)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No refresh token cookie found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Refresh token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}